Major refactor: Generate separate TOML schedule
This refactor changes how course schedules are generated and rendered.
Previously, schedules were generated and embedded directly in a
prefilled markdown file, which a user would need to open a fill out
before rendering to .docx, etc. Now, schedules are generated and
formatted as separate TOML files prior to markdown rendering. This makes
them easier to store and fill out.

+ Markdown templates are now enums
+ All variable substitution is handled by an enum method, `.render()`
+ Rename the .docx reference template downloader to prevent name clash
+ New script for generating a TOML-formatted schedule (it contains the
  `Scheduler` class)
+ TOML-formatted schedules are outputted to a separate directory, which
  the `Makefile` manages
+ Custom paragraph wrapping for multi-string TOML entries (course
  description and objectives, agenda items, etc.)